// src/openapi/model/api-user.ts
export interface ApiUser {
    /**
     * 
     * @type {number}
     * @memberof ApiUser
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    createdAt?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    updatedAt?: string;
    /**
     *
     * @type {number}
     * @memberof ApiUser
     */
    status: number;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    email: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    password?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    encryptedPassword?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    role: string;
    /**
     *
     * @type {number}
     * @memberof ApiUser
     */
    areaId?: number;
    /**
     *
     * @type {number}
     * @memberof ApiUser
     */
    experienceYears?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    selfIntroduction?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    tel?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoMemberId?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoMemberName?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoCardExpire?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    gmoCardSeq?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    gmoCardNo?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankId?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankCode?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankName?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankBranchCode?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankBranchName?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankAccountType?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankAccountNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    gmoBankAccountName?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    postalCode1?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    postalCode2?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    address1?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiUser
     */
    address2?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    emailTemporary?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    notificationCheckDate?: string;
    /**
     *
     * @type {ApiGenres}
     * @memberof ApiUser
     */
    userGenres?: ApiGenres;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    userImageUrl?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    userPublicImageUrl?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    areaName?: string;
    /**
     *
     * @type {string}
     * @memberof ApiUser
     */
    pushToken?: string;
}